+++ mytwitter/urls.py
@@ -14,16 +14,13 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.conf import settings
-from django.conf.urls.static import static
 from django.contrib import admin
-from django.conf import settings
+from django.contrib.staticfiles.urls import static
 from django.urls import path, include
 
 
-from django.contrib.staticfiles.urls import static
-
-
 urlpatterns = [
-    path('admin/', admin.site.urls),
-    path("user/", include("user.urls"))
+    path("admin/", admin.site.urls),
+    path("users/", include("user.urls")),
+    path("tweets/", include("tweets.urls")),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
