+++ user/apps.py
@@ -2,8 +2,8 @@ from django.apps import AppConfig
 
 
 class UserConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'user'
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "user"
 
     def ready(self):
         from . import signals
