+++ user/templates/user/profile/user_profile.html
@@ -17,14 +17,19 @@
         <p style="padding: 10px; font-weight: bold;">{{ page_user.user }}</p>
         {% if user.id == page_user.user.id %}
         <a class="white-btn prof-head" href="{% url 'user:signout' %}">Sign out</a>
+        {% else %}
+        <a class="black-btn prof-head" href="">Follow</a>
         {% endif %}
     </header>
+
     <div class="profile">
             <img class="prof-icon-lg" src="{{ page_user.profile.profile_img_url }}">
         <h1 style="padding-top: 50px; font-size: 50px;  float: left;">{{ page_user }}</h1>
         <div class="btn_field">
             {% if user.id == page_user.id %}
                 <a class="white-btn prof-foot" href="{% url 'user:edit_profile' pk=user.pk %}">Edit</a>
+            {% else %}
+            <a class="black-btn prof-foot" href="" style="margin-top: 100px; margin-left: 2rem; float:left;">Follow</a>
             {% endif %}
         </div>
         {% if page_user.profile.bio %}
@@ -39,6 +44,72 @@
         <h3 style="clear: both; padding-top: 10px;">Tweets</h3>
     </div>
     {% if tweets %}
+        {% for tweet in tweets %}
+        <article class="tweet">
+            <div style="width: 100%; word-break: break-all;">
+                <div>
+                    <div>
+                        <a href="{% url 'user:user_profile' pk=page_user.id %}">
+                            <img class="prof-icon" style="float:left" src='https://freesvg.org/img/abstract-user-flat-4.png'>
+                        </a>
+                        <a class="prof-name" style="float: left; margin-left: 10px; margin-top: 8px;" href="{% url 'user:user_profile' pk=page_user.id %}">{{ tweet.user }}</a>
+                    </div>
+                    <div>
+                    {% if user.id == page_user.id %}
+                        <form method="POST" action="{% url 'user:tweet_delete' pk=tweet.pk %}">
+                            {% csrf_token %}
+                                <a href="{% url 'user:tweet_delete' pk=tweet.pk %}"><button style="all:unset; float:right" onclick="return confirm('Are you sure you want to delete this tweet?')">
+                                    <img class="material-icon" src='https://cdn-icons-png.flaticon.com/512/2891/2891491.png'>
+                                </button></a>
+                        </form>
+                        <a href="{% url 'user:tweet_edit' pk=tweet.pk %}">
+                            <img class="material-icon" style="float:right" src='https://upload.wikimedia.org/wikipedia/commons/thumb/6/64/Edit_icon_%28the_Noun_Project_30184%29.svg/1024px-Edit_icon_%28the_Noun_Project_30184%29.svg.png'>
+                        </a>
+                    {%  endif %}
+                    </div>
+                </div>
+                <a style="text-decoration: none;" href="{% url 'user:tweet_detail' username=tweet.user pk=tweet.pk %}">
+                    <div class="tweet-body">
+                        <p>{{ tweet.body }}</p>
+                        {% if tweet.image %}<img src='{{ tweet.image.url }}'>{% endif %}
+                    </div>
+                </a>
+                <div>
+                    <small style="opacity: 0.4;" class="white-important d-flex full-width flex-row-reverse">
+                        {{ tweet.updated_at | date:"H:i l, y.m.d" }}
+                        {% if tweet.created_at != tweet.updated_at %}
+                        (Edited)
+                        {% endif %}
+                    </small>
+                </div>
+                <div class="tweet-footer">
+                    <div style="display: flex;">
+                        <small style="opacity: 0.8; color:rgb(0, 0, 0); margin-left:15px;">
+                            <!-- Reply Buttons Here -->
+                            <a href="">
+                                <img style="width:25px; float:left;" src="https://icons.veryicon.com/png/o/miscellaneous/basic-icon/message-54.png">
+                                <p style="float:left;">{{tweet.replies}}</p>
+                            </a>
+                        </small>
+                        <small style="opacity: 0.8; color:rgb(0, 0, 0); margin-left:15px;">
+                            <!-- Retweet Buttons Here -->
+                            <a href="">
+                                <img style="width:25px; float:left;" src="https://icons-for-free.com/iconfiles/png/512/retweet-131965017522500627.png">
+                                <p style="float:left;">{{tweet.retweets}}</p>
+                            </a>
+                        </small>
+                        <small style="opacity: 0.8; color:rgb(0, 0, 0); margin-left:15px;">
+                            <!-- Like Buttons Here -->
+                            <a href="">
+                                <img style="width:20px; float:left;" src="https://www.iconpacks.net/icons/1/free-heart-icon-492-thumb.png">
+                                <p style="float:left;">{{tweet.likes}}</p>
+                            </a>
+                        </small>
+                    </div>
+                </div>
+            </div>
+        </article>
+        {% endfor %}
     {% else %}
         <p>No tweets yet...</p>
     {% endif %}
