+++ user/views.py
@@ -8,19 +8,13 @@ from django.contrib.auth.views import (
     PasswordResetConfirmView,
     PasswordResetCompleteView,
 )
-from django.core.exceptions import PermissionDenied
-from django.shortcuts import render, redirect, get_object_or_404
-from django.views.generic import (
-    CreateView,
-    DetailView,
-    ListView,
-    UpdateView,
-    DeleteView,
-)
+from django.shortcuts import render, redirect
+from django.views.generic import DetailView
+
 from django.urls import reverse_lazy
 from extra_views import UpdateWithInlinesView, InlineFormSetFactory, SuccessMessageMixin
 
-from .models import CustomUser, Profile, Tweet
+from .models import CustomUser, Profile
 from .forms import SignupForm, PasswordForm
 
 
@@ -152,58 +146,3 @@ class EditProfileView(LoginRequiredMixin, SuccessMessageMixin, UpdateWithInlines
         if not self.user_passes_test(request):
             return redirect("user:home")
         return super().dispatch(request, *args, **kwargs)
-
-
-class HomeView(LoginRequiredMixin, ListView):
-    model = Tweet
-    template_name = "user/home.html"
-
-
-class UserProfileView(LoginRequiredMixin, DetailView):
-    model = CustomUser
-    template_name = "user/user_profile.html"
-
-    def get_context_data(self, **kwargs):
-        context = super(UserProfileView, self).get_context_data(**kwargs)
-        page_user = get_object_or_404(CustomUser, id=self.kwargs["pk"])
-        tweets = Tweet.objects.filter(user_id=page_user.id)
-        context["page_user"] = page_user
-        context["tweets"] = tweets
-        return context
-
-
-# Views for tweeting
-class TweetView(LoginRequiredMixin, CreateView):
-    model = Tweet
-    template_name = "user/tweet.html"
-    fields = ["user", "body", "image"]
-    success_url = reverse_lazy("user:home")
-
-    def form_valid(self, form):
-        form.instance.user = self.request.user
-        return super().form_valid(form)
-
-
-class TweetEditView(LoginRequiredMixin, UpdateView):
-    model = Tweet
-    template_name = "user/tweet_edit.html"
-    fields = ["body", "image"]
-    success_url = reverse_lazy("user:home")
-
-    def dispatch(self, request, *args, **kwargs):
-        handler = super().dispatch(request, *args, **kwargs)
-        user = request.user
-        tweet = self.get_object()
-        if not (tweet.user == user):
-            raise PermissionDenied
-        return handler
-
-
-class TweetDeleteView(LoginRequiredMixin, DeleteView):
-    model = Tweet
-    success_url = reverse_lazy("user:home")
-
-
-class TweetDetailView(DetailView):
-    model = Tweet
-    template_name = "user/tweet_detail.html"
